chore(overview_service): update overview whenever a new change happens
