+++ services/cd-service/pkg/repository/transformer.go
@@ -343,6 +343,11 @@ func (r *transformerRunner) Execute(t Transformer, transaction *sql.Tx) error {
 		r.Stack[idx-1] = append(r.Stack[idx-1], msg)
 	}
 	r.Stack = r.Stack[:idx]
+	if r.State.DBHandler.ShouldUseOtherTables() {
+		if err = UpdateOverviewDB(r.Context, r.State, transaction, true, true); err != nil {
+			return err
+		}
+	}
 	return nil
 }
